feat: introduce SubstraitSqlDialect
